python/requirements.txt
chainlib-eth~=0.5.0
hexathon~=0.1.7
